CSV Reader First Update

+++ modelitem.cpp
@@ -1,6 +1,6 @@
 #include "modelitem.h"
 
-ModelItem::ModelItem(const QString& date, int price, QObject *parent) :
+ModelItem::ModelItem(const QString& date, double price, QObject *parent) :
     date_(date),price_(price),QObject(parent)
 {
 }
